overlay: device_config: Disable deep press input classification
We've already disabled the touchscreen heatmap video interface used by
the TensorFlow input classifier for performance and CPU usage
reasons. However, trying to run the input classifier without the video
interface causes log spam and 3% extra CPU usage when interacting with
the device:

04-05 23:26:35.617 16591 16591 W native  : event_matcher.cc:44 Event queue exceeds maximum size; dropping oldest
04-05 23:26:36.023 16591 16591 I chatty  : uid=9999(nobody) classifier@1.0- identical 73 lines
04-05 23:26:36.028 16591 16591 W native  : event_matcher.cc:44 Event queue exceeds maximum size; dropping oldest

Logging is relatively expensive on Android, so this log spam is hurting
performance for no reason. Disable the input classifier service entirely
to fix the issue.

Signed-off-by: Omkar Chandorkar <gotenksIN@aosip.dev>