Revert "vendorsetup: Enable global ThinLTO"
This reverts commit 17f9292da24441761d56ad27102c468199c2e13f.

For unknown reasons as of now, this breaks second stage init on some
(but not all) dynamic devices. Limited troubleshooting as of now has led
nowhere. But this is not an ideal default.

For those this works well for, this is still an env flag. So use it and
disabling lto per module as you see fit.